Fix dragging behavior with touch
The nearToMouse point must be computed at/before the start of a drag. It can still be computed once per frame, but needs to be earlier. Before it was computed later in the frame, which was fine for mouse input where there's mouse movement before clicking, but not for touch, where the position is registered only as you begin a gesture.